requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# AI Services
openai

# File Processing & Image Handling
PyMuPDF
Pillow
pillow-heif
python-magic

# Document Generation
python-docx

# HTTP Client for making requests (token API, etc.)
httpx

# Environment variables
python-dotenv

# --- Testing Libraries (Optional but Recommended) ---
pytest
pytest-asyncio
